requirements.txt
numpy
gym-retro
keras
tensorflow-gpu
matplotlib
opencv-python
Pillow
